+++ NotEnoughPotions/Assets/Data/Garden/Garden 6.asset
@@ -14,4 +14,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   isPlanted: 0
   isGrown: 0
-  plant: {fileID: 11400000, guid: 3c2c555f8d9248d46b17cc6500ad62b8, type: 2}
+  plant: {fileID: 0}
