+++ NotEnoughPotions/Assets/Data/Ingredients/Cottonwood Leaf.asset
@@ -16,5 +16,6 @@ MonoBehaviour:
   description: Helps lower fevers, reduce inflammation, and ease general pain
   sellAmount: 7
   UiImage: {fileID: 8334269598078536416, guid: 54ae108c80c41584793eaa558050e522, type: 3}
+  StationUiImage: {fileID: 8334269598078536416, guid: 27cfa5159739a274382334ad9748ba66, type: 3}
   type: 0
   HarvestAmount: 2
