--- NotEnoughPotions/Assets/Data/Ingredients/Testing.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 40d3920e49b3f554b896417002080b07, type: 3}
-  m_Name: Testing
-  m_EditorClassIdentifier: 
-  ingredientName: 
-  description: 
-  sellAmount: 0
-  UiImage: {fileID: 0}
-  type: 0
-  HarvestAmount: 1
