--- NotEnoughPotions/Assets/Data/Ingredients/Testing.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 67005de741b2237408b73cbc8dc0950a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
