+++ NotEnoughPotions/Assets/Data/Ingredients/Zearalenone.asset
@@ -17,4 +17,6 @@ MonoBehaviour:
     on humans, it can disrupt hormonal balance
   sellAmount: 9
   UiImage: {fileID: 8334269598078536416, guid: 54ae108c80c41584793eaa558050e522, type: 3}
+  StationUiImage: {fileID: 8334269598078536416, guid: 27cfa5159739a274382334ad9748ba66, type: 3}
   type: 2
+  HarvestAmount: 1
