+++ NotEnoughPotions/Assets/Prefab/GardenUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 27cfa5159739a274382334ad9748ba66
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
