--- NotEnoughPotions/Assets/Scripts/ItemData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a53bec495fa451f4d901893246bd374b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
