+++ NotEnoughPotions/Assets/Scripts/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 013d6d49fa201334bbc323a20255ab2c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
