+++ NotEnoughPotions/Assets/Scripts/Items/ItemButtonUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f9b193bd35501a43b86f9fee9846382
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
