+++ NotEnoughPotions/Assets/Scripts/Potions/DisplayPotions.cs
@@ -27,7 +27,7 @@ public class DisplayPotions : MonoBehaviour
         for (int i = 0; i < stationData.Container.Count; i++)
         {
             var obj = Instantiate(stationData.Container[i].StationUiImage, Vector3.zero, Quaternion.identity, slots.transform);
-            obj.GetComponent<PotionItem>().setName(stationData.Container[i].ingredientName);
+            obj.GetComponent<ItemButtonUI>().setName(stationData.Container[i].ingredientName);
         }
     }
 }
