+++ NotEnoughPotions/Assets/Scripts/Potions/PotionData.cs
@@ -4,7 +4,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "newPotion", menuName = "Item/Potion")]
 public class PotionData : ItemData
 {
-    public GameObject StationUiImage;
     public PotionType type;
     public List<IngredientInfo> Ingredients = new List<IngredientInfo>();
