--- NotEnoughPotions/Assets/Scripts/Potions/PotionItem.cs
@@ -1,17 +1,0 @@
-using TMPro;
-using UnityEngine;
-
-public class PotionItem : MonoBehaviour
-{
-    [SerializeField] TMP_Text item;
-
-    public void setName(string item)
-    {
-        this.item.text = item;
-    }
-
-    public TMP_Text getName()
-    {
-        return item;
-    }
-}
